+++ Patches/DepositItemsDeskPatch.cs
@@ -1,4 +1,5 @@
-﻿using HarmonyLib;
+﻿using GeneralImprovements.Utilities;
+using HarmonyLib;
 
 namespace GeneralImprovements.Patches
 {
@@ -9,6 +10,13 @@ namespace GeneralImprovements.Patches
         private static void Start(DepositItemsDesk __instance)
         {
             __instance.GetComponentInChildren<InteractTrigger>().cooldownTime = 0.1f;
+        }
+
+        [HarmonyPatch(typeof(DepositItemsDesk), nameof(SellItemsClientRpc))]
+        [HarmonyPostfix]
+        private static void SellItemsClientRpc()
+        {
+            MonitorsHelper.UpdateCreditsMonitors();
         }
     }
 }